refactor: tests
It's becoming too big, "more than fingers in my hands", hence, we need to modularize this...
