+++ test/src/tests/mod.rs
@@ -1,0 +1,6 @@
+mod beta;
+mod gamma;
+mod norm;
+mod pois;
+mod rf_internals;
+mod utils;
